+++ android/src/main/java/com/hituchhimpa/reactnativesecurex/ReactNativeSecureXPackage.kt
@@ -1,4 +1,4 @@
-package com.hituchhimpa.reactnativeauthvault
+package com.hituchhimpa.reactnativesecurex
 
 import com.facebook.react.BaseReactPackage
 import com.facebook.react.bridge.NativeModule
@@ -7,10 +7,10 @@ import com.facebook.react.module.model.ReactModuleInfo
 import com.facebook.react.module.model.ReactModuleInfoProvider
 import java.util.HashMap
 
-class ReactNativeAuthVaultPackage : BaseReactPackage() {
+class ReactNativeSecureXPackage : BaseReactPackage() {
   override fun getModule(name: String, reactContext: ReactApplicationContext): NativeModule? {
-    return if (name == ReactNativeAuthVaultModule.NAME) {
-      ReactNativeAuthVaultModule(reactContext)
+    return if (name == ReactNativeSecureXModule.NAME) {
+      ReactNativeSecureXModule(reactContext)
     } else {
       null
     }
@@ -18,9 +18,9 @@ class ReactNativeAuthVaultPackage : BaseReactPackage() {
 
   override fun getReactModuleInfoProvider() = ReactModuleInfoProvider {
     mapOf(
-      ReactNativeAuthVaultModule.NAME to ReactModuleInfo(
-        name = ReactNativeAuthVaultModule.NAME,
-        className = ReactNativeAuthVaultModule.NAME,
+      ReactNativeSecureXModule.NAME to ReactModuleInfo(
+        name = ReactNativeSecureXModule.NAME,
+        className = ReactNativeSecureXModule.NAME,
         canOverrideExistingModule = false,
         needsEagerInit = false,
         isCxxModule = false,