+++ android/src/main/java/com/hituchhimpa/reactnativesecurex/SigningEngine.kt
@@ -1,4 +1,4 @@
-package com.hituchhimpa.reactnativeauthvault
+package com.hituchhimpa.reactnativesecurex
 
 import android.os.Build
 import android.security.keystore.KeyGenParameterSpec